pyproject.toml
[project]
name = "nmtfast"
version = "0.1.0"
description = "Shared library for nmtfast apps"
authors = [
    {name = "Alexander Haye", email = "ahayehaye@not-mt.dev"}
]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11,<4.0"
dependencies = [
    "pyyaml (>=6.0.2,<7.0.0)",
    "pre-commit (>=4.1.0,<5.0.0)",
    "cryptography (>=44.0.2,<45.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "argon2-cffi (>=23.1.0,<24.0.0)",
    "pytest-mock (>=3.14.0,<4.0.0)",
    "fastapi (>=0.115.12,<0.116.0)"
]
packages = [
    { include = "nmtfast", from = "src" }
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88  # Specify the line length
target-version = ['py311']  # Specify your Python version
include = '\.pyi?$'
exclude = '''  # Specify the files/dirs that should be ignored by the black formatter
/(
    \.eggs
  | \.git
  | \.mypy_cache
  | \.tox
  | \.local
  | \.pypoetry_cache
  | \.ruff_cache
  | \.venv
  | venv
  | _build
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
known_first_party = "app"
#src_paths = ["src"]
skip = [
  ".local",
  ".ruff_cache",
  ".pypoetry_cache",
  ".venv",
  "venv",
  "alembic"
]

[tool.pydoclint]
style = "google"
# TODO: figure out how to make this multi-line
exclude = '\.pre-commit.*|LICENSE|\.(lock|txt|ini|toml|md)$|^\.?venv/|\.local'
# NOTE: we are not going to duplicate types in the function and docstring
arg_type_hints_in_docstring = false

[tool.pydocstyle]
convention = "google"
# NOTE: D212 is 'Multi-line docstring summary should start at the first line' and
#   and it is mutually exclusive with D213. They both relate to the placement of the
#   summary line in a multiline docstring but enforce different styles. D212 requires
#   the summary line to be on the first line after the opening quotes, while D213
#   requires it to be on the second line. A docstring can't simultaneously adhere to
#   both styles, and this project prefers D213.
# NOTE: pydocstyle follows a strict interpretation of PEP 257, enforcing a separate
#   __init__ docstring unless explicitly configured otherwise (D107). pydoclint (with
#   Google-style docstrings) has a mutually exclusive requirement that __init__ methods
#   do not have a docstring at all. For our purposes, it's easier to resolve the
#   conflict by ignoring D107 in pydocstyle.
# NOTE: D200 wants all single line comments to have quotes on the same line, but this
#   can make documentation harder to read when putting the start and end """ on their
#   own line would improve readability
add-ignore = ["D212", "D107", "D200"]
# You can adjust the rules you want to ignore if necessary
# Only check .py files (adjust if you need more or less)
match = "^(?!test_).*\\.py"
match-dir = "^(?!alembic|\\.?venv|\\.local).*$"

[tool.mypy]
cache_dir = ".local/mypy_cache"
exclude = "^alembic/|^\\.?(venv)/|^\\.local/"
explicit_package_bases = true
mypy_path = "src"

[tool.pytest.ini_options]
cache_dir = ".local/pytest_cache"
addopts = "--import-mode=importlib"
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
# testpaths = ["tests"]

[tool.coverage.run]
data_file = ".local/coverage"
branch = true
source = ["nmtfast"]
omit = [
    "*/alembic/*",
    "*/tests/*",
]

[tool.coverage.report]
fail_under = 80
show_missing = true
omit = [
    "*/tests/*",
]
# exclude normal '# pragma: no cover' and ALL logger.* calls
exclude_lines = [
    "pragma: no cover",
    ".*logger\\.(debug|info|warning|error|critical).*\\(.*",
]

[tool.ruff]
cache-dir = ".local/ruff_cache"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.0.0"
pyright = "^1.1.396"
mypy = "^1.15.0"
types-pyyaml = "^6.0.12.20241230"
pytest-asyncio = "^0.25.3"
invoke = "^2.2.0"
ruff = "^0.11.2"
pydoclint = "^0.6.4"
pydocstyle = "^6.3.0"
black = "^25.1.0"
isort = "^6.0.1"
